refactor: everything related to multilinear polynomials in the same file
